backend/requirements/base.txt
# PatchGuard — Base Requirements
# Pinned versions from Task 1.4

Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
drf-spectacular==0.27.2
psycopg[binary]==3.2.3
redis==5.1.1
celery==5.4.0
gunicorn==23.0.0
django-redis==5.4.0
django-health-check==3.18.3
python-ldap==3.4.4
whitenoise==6.7.0
sentry-sdk==2.16.0
structlog==24.4.0
dj-database-url==2.2.0

python-dotenv>=1.0.0
tzdata>=2024.1
